+++ app/api/config.py
@@ -21,6 +21,8 @@ KI_MODEL_PATH = os.path.join(LATEST_RUN_DIR, "ki.bin")
 EN_MODEL_PATH = os.path.join(LATEST_RUN_DIR, "en.bin")
 PROJ_KI_EN_PATH = os.path.join(LATEST_RUN_DIR, "proj_ki_en.npy")
 PROJ_EN_KI_PATH = os.path.join(LATEST_RUN_DIR, "proj_en_ki.npy")
+TGT_EMBS_KI_PATH = os.path.join(LATEST_RUN_DIR, "tgt_embs_ki.npy")
+TGT_EMBS_EN_PATH = os.path.join(LATEST_RUN_DIR, "tgt_embs_en.npy")
 METRICS_JSON_PATH = os.path.join(LATEST_RUN_DIR, "evaluation_metrics.json")
 SP_MODEL_PATH = os.path.join(MODELS_DIR, "sentencepiece.model")